Fix GCC 4.6 sign-conversion warning

+++ src/stream/block.cpp
@@ -86,7 +86,7 @@ public:
 		std::streamsize nread = boost::iostreams::read(src, temp, temp_size);
 		if(nread == EOF) {
 			return false;
-		} else if(nread != sizeof(temp)) {
+		} else if(size_t(nread) != sizeof(temp)) {
 			throw block_error("unexpected block end");
 		}
 		boost::uint32_t block_crc32 = util::little_endian::load<boost::uint32_t>(temp);
